Moved all the gallery dropbox code to the client side
The server side processing that was happening before was too expensive
and redundant. We as a server have no need to see the customer's pictures.
We could if we wanted of course but we do not.
All the server side code is now using the javascript API instead of the
.net.
I also added support for the longitude and latitude displays, as gratefully
the dropbox metadata comes with that info. lots of work done for free yay!