Gallery: Added preliminary support for a widget of the dropbox folder
The pictures of our devices are not really all in one
folder but in several ones, named by date. We need to let our users
choose which flight session are we talking about so we came up with a
tile system of widgets that show the first picture of the folder and
display it. Upon clicking on the widget the folder is opened and the
thumnails and data are displayed.