Device model: Ignore User for Json serialization
We want to ignore user because the user is a complex model that
is not only easily serializable it is also too revealing of the
internals. Quite cool this kind of decorator

Also removed the json serializer settings as they are no longer necessary,
because the crazy reference loops were due to the User member
being serialized. As it is no longer serialized we simplify
the settings, and just use vanila JSon serializer.