DeviceController: When we delete a device redirect to device creation.
It seems a good fallback, that we go to device creation
after we delete one. The previous one was definitely wrong
because, to what device would we fallback when we deleted one
device? The last one available? And what if there was not other
device? You see where this is going.