Gallery: Populate thumbnails in a batch way based on a list of files
Previously we fetched the thumbnails one by one and while
we were still discovering the files. With this commit
we make a single batch request of thumbnails bases on
a list of files retrieved at file discovery time.

Also made the rendering of the blobs adapted to the
way they are available in the batch thumbnail request.