sidemenu-drohib.js: Made it more robust and beautiful
Made the side menu tree menu more robust in the parsing of
the json. This is required because we want to not destroy
the rendering of the whole page just because an error in the
side menu happened, due to for example there being no devices
associated.

Also we added a very nice plug icon to each of the devices :)