Added Map Dashboard.
Added a Map dahsboard where google maps are the backend to show
markers of where pictures were taken. Right now not much
happens in this Map dashboard besides centering around Portugal
 showing the markers of the pictures and some basic information
of where the pictures were taken. This is very rudimental and
proof of concept as the structure of the devices is not yet
taken into account and no thumbnails are shown also.