Gallery: Make Gallery display per device.
Previously the gallery was not per device but just a general list of
directories in dropbox. Given that multiple devices per user are
possible it is natural that we should have the gallery only showing
pictures respective to the device. Maps have not been updated, as
they are demo only yet.