Map.csthml: Show Exif informationf of picture
We now show the exif information of the selected
picture in the map. This information is loaded
after the thumbnail is shown, because most people
learn more information about the picture through
the thumbnail then from the EXIF, but the EXIF
can be crucial to make some inspection to the conditions
where the picture was taken.

The mechanism of content setting in the infowindow
was redone so now there is not repetition of the
content string for each of the information retrieval
stages. There is a function that generates the infowindow
html that returns html depending on the data available.