Added support for notifications through SignalR framework.
We now have support for notifications through SignalR in the
Drohub program. To achieve this we configured a hub called
/notificationshub, which shall suite to have communication
from the server to the customer in real time.

In this commit it is included the javascript libraries for the
client side loading. These libraries were generated with libman,
a tool that generates headers. Do not ask me how it works, but I
got the command to generate the js files from here:
https://docs.microsoft.com/en-us/aspnet/core/tutorials/signalr?view=aspnetcore-2.2&tabs=visual-studio-code

To actually use the SignalR js library we created a js script that
set's up and initializes the SignalR for the notificationsHub hub. We
do not do much there yet, beside putting an alert whenever we receive
a message. In this commit no code makes use of this functionality.

On the server side a NotificationsHub hub class is created with the method
send message available. It is very simple and matches the one that the client
code uses. Basically a user sends a message as a string.