DevicesController: Land and Takeoff now verify user and device id.
This is necessary so that we Land and Take off a specific associated
drone not just the hardcoded drone. It is true that at this moment the
drone address is still hardcoded but at least in this method the
id association with the model is already correctly mapped.

There is a bit of duplicated code that will need refactoring but I
need to learn a bit more about dot net to be sure on how to proceed.