DevicesController: TakeOff and Land actions emit notifications.
Now, after the action is completed on the micro-service side,
we get to notify our clients that a given device has taken off
or landed. As the code we have at this time in the client side is
very simple, the user will get an alert with status of the action
request.