Added support for Serilog logging provider.
The serilog logging provider is a bit more powerfull than the default
logging provider. One of the most important things is that it allows for
different pre-existing sinks, as well as storage of properties alongside
the logging message itself.

Also added the reference for the mssql logging sink, although no such
sink is in use at the time of this commit.

In this commit we just set up a basic console sink and the only message logged
is in the Program start. Future functionality will be added.