appsettings.json: connectionStrings no longer with SQLEXPRESS nor TrustedConnection=true
The connection strings were defined early in the project when i was
not very knowledgeable of how they worked and how they fit
with mssql server hosted in a linux container. It seems that
no only they should not have the \SQLEXPRESS nor the TrustedConnection

SQLEXPRESS namespace just seems to be plain wrong and before it was working
probably out of some fallback that was ignoring the namespace.

The TrustedConnection was even more stupid because it was enabled and
when this option is enabled Password and User Id information are ignored.