DevicesController: Refactored common takeoff and landing code
No longer do we have so much duplication in code for the TakeOff
and Landing methods. Instead we created a generic device action
dispatcher that takes the device id and the action required and
calls the appropriate service's remote procedure calls.

Also created a GetDevice method that serves as a helper for the
DoDeviceAction. It is probable that is can be useful in other
methods but for now we only use it in the DoDeviceAction.

Added some exception throws for obvious faulty cases.

I did not want to solve the problem of the dispatcher with an Enum
but instead with a delegator. But for all the coolness of c#, it is not
possible to have a static method signature of an external class to
use as a delegate parameter with a this argument. Something so trivial
in C++ can only be done with reflection, which is not type safe nor
compile time correct. Maybe i am wrong but i spent some time on it and
came up short.