DevicesController: Added Logging capabilities.
Added logging of the device rpc actions executed. May change in the
future but it now includes the device serialization of the result of the
attempt of the action (it is not an exception to fail an action..yet). It
also serializes as a properties the content of the device model.