_Main.cshtml: Initial pruning of the notification bar elements.
Lets remove the noise, both on the source as well as on the presentation. This
way we can imagine better where to go to.

Also changed the icons to be the space-shuttle. The drone icons were paid,
and i do not think that paying 99 dollars for that is worth at this moment.
Probably in the future.