Navbar: Moved the navbar out of the _Main.cshtml into a view component.
I finally have the knowledge to move the markup for the header navbar out of the
single blob cshtml. This is very cool because now i can also have a kind of global
controller that is applicable for all the page views. It also helps separate
and compartimentalize some of markup into a more logic layout.