Authentication: Log the Log-ins and logouts.
We want to register who enters and who leaves. For the login we only register
the email address of the user that logged in as the UserManager is
null on the moment the login is processed and more work would be needed to
get more info.

For the logout the whole user is logged.