Notifications: Added support for log navbar.
Added a dynamic population of the navbar with the entries of the log.
We read the log through a model called LogEntry that maps the table
created and used by the Serilog.

We got rid of the Notification model as it no longer made sense nor did
it map directly to the log table.

The view now uses the Log model directly and is a bit rough around the edges
in graphical terms but that will be fixed soon.