Added NotificationsHubPoller service.
The service polls ths database every second to check for new log entries.
If there are new log entries the entries are serialized and sent to the
client.

This was a tricky one to get because the periodic task is ran on a sigleton
scope whilst the DroHubContext runs on a scoped context. This led to some
trickery which went way beyond my knowledge of .net core. So far this seems
to be working. Beware that if you put any logging in the DoTask method
there will be an endless loop.