Finalized notifications mechanisms.
Moved all notifications.js logic into the Notifications.cshtml.
This way we can use the concept of javascript placeholders together
with the Razor place holders to have a single rendering concept for
both server side and client side code.

In this commit the notifications is populated with common javascript
that is fed json. We inject the server side json into the page through
razor but generate the view with javascript.

For when there is a dynamic update of the notifications nav bar,
the SignalR also provides a json which is fed to the same function
that generated the initial navbar.