parrot-service: First commit of parrot-service
The current state allows for RPC calls to Takeoff and to Land.
It also preliminarily prints the GPS position when the callback
cb1 is called and the message is PositionChanged. Be aware that
this requires a modification to the Olympe Drone class, to accept
callbacks to call on message received.