Logging: Changed Logging model in model and serilog.
Removed the MessageTemplate and the Properties collumn from our logs.
Instead made the SourceContext property as well as the EventId propper collumns.
The eventId should be a string but for now we can live with this. I
could not for my life, make the column a single int id.

Changed the LogEntry model to match the new way our logs are
stuctured. By default we will not want to  JSON to serialize the
SourceContext as this is internal information that is not useful
for notifications. The event Id, yes, is passed to the client.

Added a LoggingEvents helper that contains the EventId to be
passed to the Log functions.