Telemetry: Added Telemetry capabilities.
With an update of DrohubGRPCProto we are now able to
receive position telemetry. With this position from the
drone we offer the client live position SignalR messages.

We offer this telemetry through a new SignalR hub. We do not
pass the grpc telemetry directly to the client because there
are no easy way to authenticate and even if there was there
would need to be a global authenticator common with the asp.net
logic and the grpc back end one. This would make maintenance and
changes quite more fragile.

We chose to create a new Hub separate from the notifications because the
telemetry is not used only for notificvations but for other spcific
functionality like updating maps for example(not in this commit).

The telemetry received from the drone backend is a stream so it
should make the update event based instead of per polling.