Migratins: Remove migrations related to LogEntry.
The LogEntry migrations should not be handled, as they are created
by serilog itself. If Serilog does not create the table, serilog is
not able to use it, thus we lose DB logging. Also the migrations were
defective with what was the final model, so they really needed to go.