Added DronePosition model to record positions.
In this commit we update the DrohubGRPCProto submodule to
reflect the more information rich DronePosition message. As the
DronePosition generated class is very complete and a good model
to persist we then added some extra parameters as part of the partial
class extension. The extra parameter are the Device and the Id.

The Device member relates a position to a given device. There can only
be one device related to a given position entry. The Id was added as
a primary key although it does not really have much meaning. The index
is the device.

To actually record the DronePosition, the TelemetryListener records
the each received DronePosition to the database. The device is then
associated by DroHubContext through the entity framework.

To make the database creation easier the snapshot and entity migrations
were also updated.