Do not hang DroneRPC until Drone is connected.
Slightly rejecting the RAII moto we do not make the class invalid
when the drone is not connected. We try it when we initialize it
but we are ok if it fails. Instead we make the grpc class be started
and consequently the grpc server. Then we can let our grpc clients know
we are alive and exist. In the case our grpc clients ask for something
we cannot provide because the drone is not up then we have the decorator
check_drone_connected that will re-try the connection to the drone and if
it fails, we then throw an exception.

As mentioned in the previous commit there is a new smarts that tries to reconnect
in the decorator itself. This allows us to reconnect and recover from a
lost communication to the drone, whereas before we could not