Updated position message
The position message now is more complete, as specified by the new
DrohubGRPCProto. It now contains timestamp as well as device serial.
The parrot-service was updated to also provide these data.

A small refactor was done so that the queue contains the grpc DronePosition
instead of the parrot drone message. To achieve this the cb1 callback
that gets called on every message now calls a dispatchPosition that, from
a message pushes a new DronePosition to the queue.
When the getPosition grpc method is called it only needs to pop the queue
without further processing to yield the DronePosition.