NotificationsHub: Tried again to fix notifications for default cases.
Several issues fixed in such a small change. First it seems that
when you put DefaultIfEmpty the ordering goes kaboom. Due to this
we were picking the first of the id's instead of the last.

The other is that the default can also be null for whatever reasons.
Actually default is a construct for lazy sanitation and i get what i
deserve when using it.

Now the logic is taken care with clear steps not magical Or somethings.