DevicesController: Make Data use a common GetDeviceById method.
Previously we had a method called GetDevice that, with an id, retrieved
the device associated. We found out that Data controller method did not
use this function leading to code repetition.

The GetDevice function was also renamed to GetDeviceById because the argument
is an int, which is not very descriptive for type overloading purposes, so now
the method name says it is by Id explicitly.

Some logic on GetDeviceById was added to support including the positions collection