Map.cshtml: Added a sidebar to the map.
When a user click on a picture, while before only a popup
opened in the marker spot, now a sidebar opens with the same
info. The sidebar concept is much nicer visually.
Right now this is a work in progress as the side control overflows
and it cannot be untoggled.
Also the infowindows and the side-control have duplicate info.