DeviceController: Refactored DoDeviceAction to use a delegate
Finally we are using a proper concept to dispatch or shall i say
delegate the methods for the rpc calls. With a delegate that takes
a client we can have the freedom of making the rpc call we want, with
the arguments and domain specific processing without needing to
worry about the rpc connection setup and destrution. Also we
do not need a full fledged class for this which is also cool.

Adding new rpc actions, or in this case device actions, should be
more straight forward.