Initialize the grpc server only after we tried getting drone resources
It makes no sense to run the grpc calls if the resources are not
even initialized. For example in such a case where there is an
immediate grpc call but the drone object is not even created yet
a strange situation will be ran. We do not want that so we just
initialize after the drone member is inited.

Notice that we can survive any exception in drone and connection