Map.cshtml: Change container of map.
The google maps container was being clipped by other parts of the
interface, and unfortunately i never noticed before. The solution
found is to us the class provided by bootstrap. Unfortunately the
class forces an aspect ratio which sucks so much and adds a scroll.
I tried removing the padding but then nothing is shown.