Map.cshtml: Added collumn based info disposition.
Insteado of filling the screen completely with the map we are now
giving some space for extra information. This was found to be a good
compromise between simplicity and wealth of information.

In this commit only the html part is changed and templates are added.
Basically we divided the map dashboard into a grid of 2/12 data and
10/12ths map.