Map.cshtml: Add a populatePictureStack method to the populateExifInfo.
In this commit we add a box with the picture info and thumnail
to the Picture Stack box. This is not very clever yet as the collapse
functionality is basically still broken. Also it may happen that
a duplicate box is added which is kind of stupid. In the future
we should try to check if the id of a given picture is already
in the stack and just expand it if the user requests it.