Map.cshtml: Add a mock of the Devices box content.
In this commit we add the mocking of design that will have information
about the device while in the map dashboard.
The data is populated with the ViewBag Device List and razor pages construction.

The takeoff landing and movetoposition actions are now available per device
and with the correct action url provided by the devices iterated from the DeviceList.

The GotoPosition toggles a box where the user will input the coordinates
where to move the drone.