Map.cshtml: Added support for right click on the map move to position.
This commit adds the logic and markup that adds a modal that allows to select
a device, for which the gathered coordinates apply. After all device is selected
on the modal, the MoveToPosition box expanded and is populated with the coordinates
from the right click.

The altitude input is focused as, this data needs to be input by the user.