Map.cshtml: populateMarkers now takes a serial and id parameter.
populateMarkers now has a serial and id field. The reason for the id
is so that we can unambiguously identify a marker.
The reason for the serial is so that all the markers have a related serial
number associated. I am not completely sure about this parameter but
it makes it easy to group markers related to a given device. The device serial
number cannot serve as an identifier because there might be many markers
related to a given device serial number.