Map.cshtml Added functionality center on device.
Each device box now has a map icon which will center the map
on the device. A global register of markers in our page is
created through global variable. This is required because
it is not possible to query google maps for available markers.
At least as far as I know.