Added check and ballances to the Drone object.
Created a family of Drone object wrappers that
when chained allow for transparent control over
the singleton drone object.

In this commit they are DroneRAII, DronePersistentConnection
and DroneThreadSafe.

DroneRAII makes sure the connection is released when
the drone object is destroyed. There is a funny
anti-pattern here because there is the possibility
the _drone object does not exist at the time the del
operation is called.

DroneThreadSafe makes the _drone object be accessible
only through a lock.

DronePersistenConnection makes sure there is an observer
polling whether we are connected to a drone or not,
and retry when applicable