Map.cshtml: Added value classes to device information.
This is an experiment where we assign classes instead of id to elements.
The goal is that we can have a piece of code update any element
text that contains the class. This way we do not need to track down
thousands of unique identifiers.

To the above mentioned end the device information lines on the side box
now contain this magic classes, with the disambiguation of the serial number.