Added support for logging instead of printing.
As the micro service will eventually run for a long time and
the printing information will probably need to be processed in
different scenarios it makes sense to use the logging facilities
instead of prints.
Furthermore we now get easily the ability to turn on verbosity,
which is triggered depending on the command line arguments passed.