Map.cshtml: Get the last position of a given device.
With this commit we add support to pass the last position recorded
of a given device. This way even if we did not yet get a real position
we can already put a marker where the drone would be expected.

Moved the code to put the marker out of the telemetry part and put it
outside in a way that it can be re-used by the telemetry as well as
the initial static page loading.

To keep the data format representing the DronePosition consistent between
the static page population and the dynamic page population, each device
has a <script> clause with json in the static page.