Split Drone arguments in keys wheter we are a simulator or real thing.
The SkyController connection did not make sense in a simulator and
hung the sim. So we only enable the SkyController in the real drone.
This required some refactor of how the arguments are passed around the
multiple drone classes, to make use of the kwargs way.

Effectively we moved the configuration of the Drone object to the
respective classes.