Added DeviceMicroServiceOptions and respective appSettings.
In the future the location of where the drone service is located is
going to be configurable. This is of course not the way it should be
done correctly, but it allows us some degree of configuration for
the developer, deployment scenarios. In the future these options
should be associated to the Device models.