DroHub.csproj: Force bower_components to be put in the deployment.
For some reason dotnet actively removes bower_components directory
from release deployments. This is very unfortunate because we have
several links to files in that directory. I do not have time to go
find what is wrong with it, and bower seems to not be very happy
in my development machine so this is why i did not go further.
These javascript hippies...