Prepare application to be served through airborneprojects backend.
In this commit we remove the possibility of serving any https
connetions as this is to be done by the front end server of
airborne projects.
We also adapted infrastructure addresses for the case where the
application is run in release mode. Specifically it needs to
accept and add headers for the forwarding of the http connection.