parrot-backend: Finished migrating parrot-backend to this repository
The previous 27 commits were commits that were transplanted from
another repository that we are now merging in this.
Because of this we now merge the gitignore and delete the submodule
accounting done for DroHubGRPC as it is already part of the repository
of DroHub to which were are now transplanting to.