Added Support for RadioSignal.
This commit introduces the whole stack of propagating the RadioSignal
all the way from the drone backend to the browser.

The model for the DroneSignal is not connected to the database
yet. This needs to be done.

The signal quality does not do much on the web browser yet, but
will probably be connected to some color cue.