Map.cshtml: Make the default data label background default.
This way we are able to have the default color as grey.
When data is received they become blue as the primary
color.

Right now as the only data we have connected to the
device dashboard is the telemetry signal and the battery
only these 2 data types are able to become non default.