parrot-service: Remove hardcoded serial and use real one.
Up to now the serial number was hardcoded and faked. With this
commit the containers are passed the serial number object and
the serial number is retrieved from this object everytime it is
required.

One important detail about this new way of getting the serial number
is that it may fail the container dispatch function that requires the
serial number. This is on purpose. We do not want to forward to our
users any telemetry which is not trackable to a device.