FlyingState Added support for FlyingState telemetry. Fullstack
This commit contains changes for the whole implementation of FlyinState.
It has adaptations in the parrot-service, the dotnet server as well as in
the Map.cshtml.

One special note is about the fact that we change the color and maniuplate
the plane icon of each of the devices according to the flying state we
are currently.

Includes database recording.