All telemetry models inherit new IDroneTelemetry.
With the new interface IDroneTelemetry we are able to enforce that
every telemetry message class needs to have serial number and timestamp.

This also led to a modification on the RecordTelemetry method where now
it no longer is a generic method but instead a method that accepts an
ITelemetry instance.