TelemetryHub: Block on tasks to complete.
We had a broken assumption that we could just abandon the tasks for
the telemetry gathering, but this is not really a good assumption.
For the case where they are canceled abnormally we can have some status
and logging. The same for the other case. Before this was not possible,
as a design