TelemetryHub: Do not save information of devices which we are unaware of.
Previously, if we received a telemetry for a device which was never
registred, then we would have a crash saving the data. After all
how can we save data about a device we know nothing about. Now we
check for whether the serial number is known. If it is not then we
just discard it.