DeviceRepositoryController: Only include the last of each telemetry on the GetDeviceListInternal
Dumping the whole battery logs and other telemetries is kind of insane,
and will have performance impacts in deployment. This way we
just iterate over the devices associated with the user and add the
last entries for the telemetry of each device. Probably there is a more
efficient way of doing this, with less DB roundtrips but for now I
do not want to spend time on this. This can be very easily refactored in
the future.