Added possibility for an action to request the file list.
The file list on the drone mass storage device is very loosely
defined, and the identifiers do not need to mean much for whoever
is reading them. They are identifiers though, and if in the future
the FileEntry of a DroneFileList is requested from the Drone, the
drone should be able to know how to retrieve it.

That part is not yet implemented.