Added ability to have telemetry of new file lists.
This is important so that we can trigger actions on the backend
for, for example, to upload files into a given cloud storage.
There is no recording of this telemetry yet, because i do not see
a direct need. Also this information is not displayed yet due
to me not having figured out yet how to present this information.