DeviceMicroService: Make the telemetry tasks a member variable.
It has come the time to make the telemetry_tasks list be a member variable.
The reasons is that we no longer expect tasks to stay awake just waiting
for telemetry, but also we expect to have tasks that will do assynchronous
device actions. To this end we need to track this tasks in a more global
way.