DeviceMicroService: Add possibility to cancel action tasks.
Actions are not implemented in this commit but they are basically tasks
that request some kind of functionality from the drone backend. As they
are requested on demand they are not executed through the ExecuteAsync.
As they are not called through the ExecuteAsync they will not share the
CancelationToken. As i do not want to invert the class in a way that it
is only constructed through the ExecuteAsync, i created a parallel
CancelationToken that gets canceled when the stopping token is itself
canceled. This way I am able to propagate the background service's
cancelation to any action task that might be running.