Moved Device Actions to DeviceMicroService.
This way we can have the controller not know anything nor need to
manage grpc connections to the Device grpc server.
The commit diff looks big but it is mostly just shufling code.
Also now the action code is Async by default, and callers
can await on it if desired. An example is that for the GetFileLit
the result of the action is awaited but Takeoff is not.

To make the DeviceMicroServices accessible to the controller
we needed to launch it in a slightly different way, otherwise
it would not show up in the dependency injection candidates.
Stolen from
https://stackoverflow.com/questions/51254053/how-to-inject-a-reference-to-a-specific-ihostedservice-implementation