Improved logging setup.
Previously we were setting logging in a way that confused dotnet
and it logged things twice. Also the logging was only initialized
in the Startup class instead of in the Program itself. Another
improvement is that the log now flushes when the program is closed.

This should improve the quality of the logging and confort.