Added JanusService.
We needed to add a JanusService REST API helper so that we can manage
the available ports to receive video. The current API allows for completely
unmanaged video resource allocation, so it allows for creation of a session,
handle, create an rtp mount, detruction and listing.

I mean unmanaged in the sense that session and handles and mountpoints
are not tracked.