JanusService: Added GetStreamInfo and getAvailableMountPointPorts.
Due to the fact that the stream list request does not return all the
info of each stream we need to further get the StreamInfo of each
stream.

With this information we can query the system about what are the available
ports. This information can be useful when you are about to add a new port
and want to add a port which is available.