Factored out Video stuff into a ViewComponent.
The Video presentation was moved out of the Map.cshtml because
we are going to convert the javascript logic to the server side,
so we need a place to put this logic. A ViewComponent is the right
way to to this, as the video presentation can be included in many
different view scenarios.