JanusService: CreateSession and createSession refactoring.
CreateSession objects now populate themselves without any
input from the exterior, as a data structure ready to be
used to create a new session. No more needs the createSession
to craft a CreateSession object. Just creating it will be
enough.

Also we now are trying a new method getJanusAnswer which tries
to factor out the connection handling part of the createSession
method. In the future this method should be used by all the
Janus communicating methods.