JanusService Renamed and Refactored createHandle to createStreamerPluginHandle
Similar refactoring as what was done the createSession. Use
getJanusAnswer and moved logic to the data model instead of
doing it in the action itself.

The rename was due to the fact that createHandle being non-descript
and we need to make the user aware that the handle is for the
streamer plugin not any janus general handle.