JanusService overhauled the createMountPointMethod.
Key differences is that it no longer takes implementation specific data structures,
and instead it takes a list of arguments relevant for the setting up of a video
session. This is in contrast to the previous situation where the user of the method
was responsible for the arbitrary population of mountpoint information.
No longer does the user of the method need to assign for example ports nor guess
what is required.

DUe to this change in paradigm we now return a datastructure specifically for public
consumption with the data about the created mountpoint.

Another key change is that JanusRequest.RTPMountPointInfo now takes an id. This id
is a key for the stream. Previously this was not passed in the request and this resulted
in Janus generating it's own Id for the stream, which is quite a stupid lost of
opportunity for us.

Finally we also try to catch the errors returned in the JanusAnswer which are specific
to the creation of a mountpoint.