JanusService lock down encapsulation to internal.
As much as is possible in a class with no state, we set most of the
models and submodules to internal, so they are not visible outside
the assembly. This way we can better control the interfaces with the
outside world, and not suffer breakages when for some reason Janus
changes it's REST API.

To achieve this we needed to move JanusStreamListInfo out of its nested
structure to a top level and removing the Janus prefix, as it is now
a non janus specific data structure.

For some reason the compiler is now saying that JanusStreamingPluginData is now
unaccessible but this is not true because the JanusAnswer JSON deserializer
populates this data structure. We silected it accordingly.

Also removed getAvailableMountPointPorts because it did not make sense
anymore because the mountpoint ports are not useful, the URIs are.