DeviceMicroService: Added support to create rtp mountpoints on demand.
Whenever we have a connection to the remote drone micro service,
we create a janus rtp mountpoint and tell the drone micro service
to start streaming the video to this mountpoint. The error
handling is very crude at this point but it is working.

Also the device gets the mount point information associated with
it so that the users of the device model can be aware of it.