Added support for HeartbeatMonitoring to ping the service.
The first part of this change is related to the fact that we now
have a pingService rpc call that we can use to verify that the
drone and the remote service are available. The semantics is that
a NOK will mean that the drone service is reachable but the drone
itself is not. On the other hand a pingService throwing an exception
normally means that the drone service is offline or not running.

The second part of the change is that instead of having all the telemetry
trying on it's own to connect the drone service, it will only do so
when the heartbeat service gives the ok. This saves resources and spams
less the logs and resources for the case where the connection is down.
Also in sequence with this heartbeat monitoring philosophy when the
heartbeat fails, then all the telemetry tasks are canceled and will
only be respawned on a successful connection. This effectively centralizes
the connection management.

The last point is that the monitor spawns heartbeats monitors per device
instead of having a single possible device connected per server instance.
Note that this is incomplete because in the current state it is assumed that
all devices are accessible through a single configurable address. It is
expected that in the future this will not be the case.