Added and make use of serial and timestamp for DroneReply.
The pingService and heartbeat monitors relied on DroneReply but DroneReply
did not contain the Serial number of the Drone which could have led to a
DroneService serving multiple devices Id/Serial when in fact it only one was
relevant. With the serial we can not verify that heartbeat is indeed from the
device being monitored. A case where this is possible is with the case where
I have a simulator and a real drone reachable through my machine. Both have
the Drone service available in my machine but this Drone service can be one of
each.