parrot-service: pollProcess is a generator in itself and improved flow control.
The new concept makes the pollProcess a generator that the rpc method uses
to generate messages. It also cancels the context after it is verified
that the video stream that was originally requested has died or any problem
happened. This way we can start the cleanup faster.

Speaking of which we removed the cleanupProcesses in favor of a cleanProcess
that just cleans a specific process and it's resources.

Note that when the context is done the cleanup is automatically ran. This is for
the situation where nobody is calling our generator anymore due to the connection
being canceled, but where we want an opportunity to clean the process of the encoder.