parrot-service: Remove locking in the ParrotSerialNumber class
It turns out this is a very heavy requirement, which gets loaded onto
every serial read. This is extremely pessimistic because the first information
we get is the serial. Furthermore the whole code is made to expect a failure
on serial get. Also in Python there are no out of order execution and we are
guarding the read of the serial with a _set variable. So at worse we miss a
successful Get call when we have the serial already stored but the _set variable
has not yet been toggled. Very cheap price to pay, compared to whole program
contention with the lock.