Video frontend: Modified partial Video view and adapted streamingtest accordingly.
This commit is still on the road of narrowing down the functionality of
streamingtest.js that came from the original janus javascript into fitting
what we decided in the view.

The basic usage at this point is that the partial view provides a box related
to the device Id. It has a play button that starts the video, and expands the box
and it has a stop button that just stops the video and displays a message saying
that if you need you need to press play. Maybe a bit idiotic but the goal is
simplicity of functionality at this moment. We still cannot have 2 videos playing
at the same time. We are able to correctly play a video though