DroneMicroService: Added a retry for the ExecuteAsync method.
I thought previosly that was no situation where the spawnHeartBeatMonitors
could abort besides a cancelation token but it turns out that there is.
If there are no devices on the server then the code just throws an exception
and will not retry. This is not how this service is supposed to work so if we
get any exception we catch it and we only let the flow go out of the method if
we get a cancelation token.